# docs/swagger.yaml
basePath: /
definitions:
...
  vo.Result:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 响应数据
      msg:
        description: 错误信息
        type: string
      requestId:
        description: 请求ID
      timeStamp:
        description: 响应时间戳
    type: object
host: localhost:9010
info:
  contact: {}
  description: This is the API documentation for Jank Blog.
  version: "1.0"
paths:
...
  /verification/sendEmailVerificationCode:
    get:
      consumes:
      - application/json
      description: 向指定邮箱发送验证码，验证码有效期为3分钟
      parameters:
      - description: 邮箱地址，用于发送验证码
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 邮箱验证码发送成功, 请注意查收邮件
          schema:
            $ref: '#/definitions/vo.Result'
        "400":
          description: 请求参数错误，邮箱地址为空
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误，邮箱验证码发送失败
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 发送邮箱验证码
      tags:
      - 账户
  /verification/sendImgVerificationCode:
    get:
      consumes:
      - application/json
      description: 生成单个图形验证码并将其返回为Base64编码字符串，用户可以用该验证码进行校验。
      parameters:
      - description: 邮箱地址，用于生成验证码
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回验证码的Base64编码
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: 请求参数错误，邮箱地址为空
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: 服务器错误，生成验证码失败
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
      summary: 生成图形验证码并返回Base64编码
      tags:
      - 账户
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"